# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi",
        "description": "项目前端api",
        "contact": {
            "email": "953068028@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://local.text.com/api/v1",
            "description": "测试服务器"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get project information",
                "description": "用户注册",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "verifyCode",
                        "in": "query",
                        "description": "验证码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nickName",
                        "in": "query",
                        "description": "昵称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "用户手机号",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "用户密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{'status':true,'status_code':0,'message':'','data':{'mobile':'18228068391','nickname':'1111','avatar':'http:\\/\\/admin.17dushu.com\\/vendor\\/presets\\/avatar.png','updated_at':'2020-05-12T12:17:52.000000Z','created_at':'2020-05-12T12:17:52.000000Z','id':2}}"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get project information",
                "description": "用户登录",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "用户手机号",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "用户密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get project information",
                "description": "用户获取个人信息",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://local.text.com/oauth/authorize",
                        "tokenUrl": "http://local.text.com/oauth/token",
                        "refreshUrl": "http://local.text.com/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}